Add support for key expiration
This commit implements support for `EX|PX|KEEPTTL` expiration options
and few other options, such as `GET` and `NX|XX`. Note that options
`EXAT|PXAT` are accepted, but not implemented in the storage layer.

It's also importation to note that this commit contains multiple
refactorings and support for (de)serialization of some new formats:
integers and simple errors.