fix(repl): send GETACK to all replicas during WAIT
This commit changes the way replicas are requested to send their offset
acks as follows:
  1. We determine (without sending any requests) whether there is _at
     least one_ replica that is behind (if not, then we're done)
  2. If so, we then send GETACK to _all_ replicas and await their acks,
     updating local views

The issue with the previous implementation was that in step 2 we only
requested those replicas that were deemed behind. This caused an offset
mismatch when only certain replicas got the GETACK request and update
their offset, but those that were previously up to date didn't and thus
got behind for no reason.