feat(rdb): implement RDB writer
This commit adds an implementation of an in-memory RDB writer and uses
it to write a snapshot of current store during PSYNC (in the leader).

Additionally, replicas no longer read the dummy empty RDB file after
handshake and rather load the RDB received from the leader into their
initial store.

Known issues & limitations:
 - BGSAVE is not yet implemented, so the server does not actually to a
   process fork to persist its store state to the FS (as an RDB).
   Instead it just clones the store and serializes it in memory.
 - Similarly, replicas do not overwrite their local dump file upon
   initialization.
 - Also there might (as a result) be some unnecessary allocations and/or
   in-memory materializations.