feat(transactions): implement transaction EXEC
This commit implements non-replicated transaction execution after
receiving an `EXEC` command.

Note that in order to support EXEC, the `Store` has been refactored, so
that it's possible to execute multiple commands while holding an
exclusive lock to the database/store.

The EXEC command execution takes an exclusive lock to:
 - A database if all buffered commands execute against client's current
   database
 - The entire store, if there's at least one SELECT command with a
   database other than the current one

Known issues & limitations:
 - Commands queued up and executed by an EXEC are not replicated
 - Nested transactions are not supported